Create running_sum.c
This commit will create the mathematical algorithm for the running sum
of 5 positive integers inputs. I used "while loop" and if-else
conditional statements. Also, I intentionally used the "goto" loop
execution flow control statement inside the block scope of the "else"
statement in the "while loop" just to test its functionality and
performance with the algorithm.