Refactor running_sum.c
This commit will refactor the mathematical algorithm for the running sum
of 5 positive integers inputs using a different style from the previous
commit. I used "for loop" and if-else conditional statements. Also, I
intentionally used the "break" loop execution flow control statement
inside the block scope of the "else" statement in the "for loop" just to
test its functionality and performance with the algorithm.